feat: Improved exif date parsing. fix: Fixed bug which started the picture list over from the beginning even though it didn't change. fix: Fixed bug which reloads pictures wether db update has occured or not. fix: Fixed increment of times_shown for selected pictures. fix: Fixed bug in get_next_picture leading to infinite recursion. fix: bug in info logging for db update. fix: Fixed logging for adding/updating pics to db. fix: small bug in get_next_picture func fix: Fixed bug where the last picture in the loaded list does not exist. fix: Change db update/clean sequence to catch moved pics before they are deleted fix: Fixed rating filters. fix: Error while handling problems with db updates of pictures. fix: Fixed 2nd bug in rating filter. log: improved logging for next picture manager. log: Improved logging around reloading pictures. log: Improved logging in PictureUpdater. log: Improved logging for deletion of removed pics from db. build: Temp fix for documentation builder. test: Fixed tests. test: Improved tests by utilizing in memory sqlite db.

+++ pipictureframe/picdb/DbObjects.py
@@ -8,7 +8,7 @@ class PictureData:
         self,
         hash_id: str,
         abs_path: str,
-        mtime: int,
+        mtime: float,
         orig_datetime: datetime,
         orientation: int,
         rating: int,
